fix playing

+++ src/core/src/club/history.rs
@@ -4,13 +4,13 @@ use chrono::NaiveDateTime;
 pub struct MatchHistory{
     date: NaiveDateTime,
     rival_team_id: u32,
-    score: (u8, u8)
+    score: (i32, i32)
 }
 
 impl MatchHistory {
     pub fn new( date: NaiveDateTime,
                 rival_team_id: u32,
-                score: (u8, u8)) -> Self {
+                score: (i32, i32)) -> Self {
         MatchHistory {
             date,
             rival_team_id,
