+++ src/core/src/league/result.rs
@@ -50,8 +50,8 @@ pub struct LeagueMatch {
 }
 
 pub struct LeagueMatchResultResult {
-    pub home_goals: u8,
-    pub away_goals: u8,
+    pub home_goals: i32,
+    pub away_goals: i32
 }
 
 impl From<ScheduleItem> for LeagueMatch {
